Add support for Nordicbits.org/nordicb.org (#3552)
* Create Nordicbits.cs

Add support for Nordicbits / DK Private Tracker

* Create ConfigurationDataNordicbits.cs

Create Nordicbits.cs

Add support for Nordicbits / DK Private Tracker

* Update README.md

Added a new tracker Nordicbits

* Update Nordicbits.cs

The self-tester didn't work, because of the trackers use of 'Today' and 'Yesterday' in releases. Where other releases are "MMM dd yyyy".
This is fixed now.

* Update Nordicbits.cs

Search string wasn't build proberly, due to the catagory of nordicbits.
The images, which we use to determin which catagory the release is from goes like this - 
cat=5
But when we need to search under that catagory and have more than one, it should go like this -
cats2[]=5

Changed the categorieslist, by using replace and changed the query string to match the original search string.

* Update Nordicbits.cs

Should solve the bug - The string was not recognized as a valid DateTime. There is an unknown word starting at index 0.

* Update Nordicbits.cs

Tracker don't support imdb, therefore problems with handling of couchpotato, radarr or similar.
Fixed by setting imdb to false;