torrentcouch: added this public tv tracker. resolves #3797
this site sucks.
the search results contain just the title and details link, no files, seeders, date or other useful details.
and despite search results specifying 720p for example, the details page contains download links for any related quality of the show, its pot luck which is the first  link, it might be SD, 720p, .torrent or magnet, or antyhing else for that matter.
the site is not really suitable for jackett but ive added it for those that want to use it.