Resolve Issue with Download Volume Factor
I apologize for my inexperience, but I was able to resolve an issue, so wanted to share. I noticed that the downloadvolumefactor was not detecting freeleech items properly. From looking at the html through chrome, the span should contain Freeleech just like the existing code was written, but that wasn't working. I enabled enhanced logging, and when looking through the dump from enhanced logging, noticed it should be FreeLeech. I changed the code locally and it resolves the issue.