Add crazyhd.yml
This was previously this was closed as needing c#, but I disagree. I salvaged the attempted yml file and modified it with the fix. I did not change anything from the original except the search block. The original issue for this tracker was that it didn't support the search query as a get parameter, but it does as a post. Also updated the freeleech text replacement to be less strict.