Update BJShare.cs to receive year for animes
In the current version, BJShare doesn't append the collected year to animes parsed. This causes mismatches in Sonarr and Radarr because they use the year of the show when parsing the data. This causes Anime that are reboots (like Digimon Adventure 2020) to not be identified and anime movies (that are in the anime section) to not be correctly identified.