gaytorrentsru: update download
some users are restricted in the number of downloads they are allowed  per day.
when a dl for a torrent is tried the first time, an intermediate page is presented with the available dl slots and a query to continue.
subsequent DL for the same torrent does not get the intermediate page.
This patch bypasses the query, but if users exceed their daily slots the download will fail, as it should.