Update bitsexy.yml
bitsexy - Site's 'search help' lies, prepending with `+` does not make a word required, only preferred (which is supposed to be the default search, but isn't). That's still useful to us, as it moves matches to the top of the first page, but andmatch is still required - surrounding words with `"` is ignored, surrounding whole phrase with `"` uses 'whole phrase' search, replacing whitespaces with `.` is ignored. Site searches both titles and descriptions, no option to change this.