Update azure-pipelines.yml
this causes the system to wait for an existing azure-pipeline to complete before submitting a new one. 
while a pipeline is active all pushes are batched together and a new run is started.
this should prevent  multiple pipeline clashes, and reduce the number of runs conserving our free allotted runtime quota.