Yify sites, proxies, status links updated
As per the yify status & proxies website, these are the only ones which are correct YTS sites. I though have kept the yts.unblockit.dad as it is the only other one which seems to be leading to the same certificates as the other yify sites.
Moved rest all to legacy links section.
Also, added a section for status & proxy links for future referencing.
Please let me know or place it correctly in the file as needed, but keep those 2 links there (status & proxies).