fix(freefarm): Captcha/flaresolverr fail -> cookie
For some time, the Prowlarr edit form for this definition fails to load the
captcha. Trying to save the form fails with:

    HTTP request failed: [500:InternalServerError] [POST] at
    [http://flaresolverr:8191/v1]

This happens whether or not the indexer is tagged with the tag on
"FlareSolverr" under "Indexer Proxies". Regardless, reverting to
cookie authentication allows the edit form to be saved successfully
and the resulting indexer also then successfully returns results in
the Prowlarr search UI.