feat(hdbao): Add new HDBao Chinese private tracker
This is my second new indexer definition. Thanks to the maintainers'
gracious response to my previous request for education about what to
test, I know much more what I don't know. So this time I had more
questions while writing the definition before submitting a PR and I
captured them in `# TODO: ` comments. Please reply to each of the
questions in each of those comments in your review. As I ingest review
feedback I'll cleanup those comments and force push without them.

For this new definition, I performed all the testing I originally did
for my previour PR plus all the additional testing I learned from the
maintainers on my previous PR. I'll do the same as I incorporate
feedback for this new definition.

This PR includes this commit and the previous which just copies the
example definition I started from. Its valuable for VCS history to
show which definition a new definition is based on, but it's an even
more valuable reference about the differences between indexers for VCS
history to show the *changes* made to the example definition to create
the new one. Subsequent changes from review feedback should be
squashed into this commit, but not into the previous commit so the
changes from the example definition are preserved in VCS history.