feat(letseed): Add LetSeed private tracker
This is my third new indexer definition but still includes new territory for me:

- My first `download:/before:/path: takethanks.php` indexer
- The only PHP indexer I can remember that embeds the torrent DB IDs into the details
  and download URLs as opposed to query parameters

So pay particular attention to those factors when reviewing.

There are also still some `# TODO: ` comments from me to address. One is specific to
this indexer and the others are general indexer definition questions that hadn't
occurred to me before.

Fixes #15901